requirements.txt
yfinance==0.2.36
anthropic==0.25.0
supabase==2.4.2
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
pytz==2024.1
